handwriting_text_recognition_AI_project/requirements.txt
torch==2.8.0
torchvision==0.23.0
transformers==4.57.3
Pillow==11.3.0
Flask==3.1.2
requests==2.32.5
easyocr==1.7.2
pytesseract==0.3.13
